pyproject.toml
[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[google-genai,huggingface,ollama,openai]>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "langchain-weaviate>=0.0.6",
    "langchain-nebius>=0.1.3",
    "python-dotenv>=1.2.2",
    "langsmith>=0.7.16",
    "langgraph>=1.0.10",
    "redis>=7.3.0",
    "psycopg2-binary>=2.9.11",
    "boto3>=1.42.64",
    "requests>=2.32.5",
    "httpx>=0.28.1",
    "unstructured[local-inference]>=0.21.5",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "markdown>=3.10.2",
    "beautifulsoup4>=4.14.3",
    "tiktoken>=0.12.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "typer>=0.24.1",
    "aiohttp>=3.13.3",
    "tenacity>=9.1.4",
    "orjson>=3.11.7",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
    "cryptography>=46.0.5",
    "passlib[bcrypt]>=1.7.4",
]
